requirements.txt
numpy
cv2
pyserial
odrive
pyrealsense2
pyglet
pygame